requirements.txt
# requirements.txt

playwright
gspread
google-auth
beautifulsoup4
requests
tenacity
